[type-inserter] Traverse the AST, insert placeholder type ("any") for now
Only the relevant nodes (function declarations, parameters, and variable
declarations) are updated. There are some special cases to avoid, e.g.
for...in and for...of loops contain declarations, but these do not
support type annotations. Also, binding patterns (or destructuring
patterns) cannot be annotated (the library crashes if you try).

The next step is to read in a CSV file of types, and then update the
corresponding identifiers.