[wip] More hacking away at type-inserter
This version "mostly" works to insert types properly. It does not handle
arrow functions `(a, b) => a + b`, and the algorithm needs to be
modified to support that.

However, it works for variable declarations, function declarations
(parameters and return types), and function expressions (parameters,
return types if the expression is named).

The code is also very messy and repetitive.