Remove the data and scripts for running DeepTyper with the JS lexer
Using the TS lexer can cause some problems with tokenizing JavaScript;
in the npm-nodeps dataset (which has 2500 JavaScript files), only one
file was handled incorrectly.

This is a small enough number that there's little reason to modify
DeepTyper just to handle that case.

This also reverts the commit

"[DeepTyper] Make it easier to switch between TS and JS lexers"
3d34177a55e67fa4eff3a592341b58180f74a231.