Investigate some of the type-inserter warnings
The problem is that type-inserter does not match a token sequence if the
preceding token is "for"; this is to avoid matching declarations in for
statements. However, in this case, the "for" token is actually a
property name, not a keyword.

There is no good way to avoid this situation, but it is rare (only 4
occurrences in 2 files, out of ~2500 files).