Optimize migration script and Incoder
Instead of a startup cost for each file or project, the migration script
starts Incoder once, and passes it the entire list of directories.
Incoder will loop over the JavaScript files in each directory.

At the same time, the migration script polls the filesystem for a "done"
file, so it knows when inference is complete for a package.

This is the same strategy for handling LambdaNet's startup cost.